// frontend/projects/modeling/src/lib/api/generated/kgs/model/ontologyInfo.ts
/**
 * PLASMA KGS Service
 * Service for maintaining the Knowledge Graph
 *
 * The version of the OpenAPI document: 1.4.0
 * Contact: pomp@uni-wuppertal.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OntologyInfo { 
    prefix?: string;
    uri?: string;
    label?: string;
    local?: boolean;
    description?: string;
    noProperties?: number;
    noClasses?: number;
    classesCount?: number;
    propertiesCount?: number;
}